backend/requirements.txt
Django==5.2.8
djangorestframework==3.14.0
django-cors-headers==4.3.1
PyMySQL==1.1.0
cryptography==41.0.7
Pillow==11.0.0
python-decouple==3.8
celery==5.3.4
redis==5.0.1
requests==2.31.0
